requirements.txt
Django>=5.0,<6
djangorestframework>=3.14
djangorestframework-simplejwt>=7.3
drf-spectacular>=0.27
django-redis>=5.4
redis>=5.0
requests>=2.31
django-environ>=0.11
django-cors-headers>=4.3
gunicorn>=21.0
